Refactor : 패키지 설정
