--- BE/src/main/java/com/codesquad/todo8/WebConfig.java
@@ -1,22 +1,0 @@
-package com.codesquad.todo8;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-  private final JwtInterceptor jwtInterceptor;
-
-  public WebConfig(JwtInterceptor jwtInterceptor) {
-    this.jwtInterceptor = jwtInterceptor;
-  }
-
-  @Override
-  public void addInterceptors(InterceptorRegistry registry) {
-    registry.addInterceptor(jwtInterceptor)
-        .addPathPatterns("/api/**");
-
-  }
-}
