+++ BE/src/main/java/com/codesquad/todo8/controller/TodoRestController.java
@@ -1,14 +1,14 @@
-package com.codesquad.todo8;
+package com.codesquad.todo8.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("api")
+@RequestMapping("/")
 public class TodoRestController {
 
-  @GetMapping
+  @GetMapping("hello")
   public String hello() {
     return "hello";
   }
